利用cmake模块化构建项目
