--- clean.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-cd /home/gzf/cjml_server || exit
-rm -rfv bin CMakeFiles cmake_install.cmake CMakeCache.txt Makefile
